requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.3.1
sqlalchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
bcrypt==4.1.1
pytest==7.4.3
pytest-cov==4.1.0
